refactor: moved constants to diff. file
