+++ src/constants/index.tsx
@@ -1,6 +1,5 @@
 import React from "react"
 
-const tagline = "API Platform"
 export const githubRepoURL = "https://github.com/tailcallhq/tailcall"
 export const tailCallBenchmarkUrl = "https://github.com/tailcallhq/graphql-benchmarks#benchmark-results"
 export const tailCallBlogUrl = "https://blog.tailcall.run/"
@@ -528,20 +527,4 @@ export enum Theme {
   Light = "light",
   Dark = "dark",
   Gray = "gray",
-}
-
-export const PageTitle = {
-  HOME: tagline,
-  ABOUT: `About | ${tagline}`,
-  ENTERPRISE: `Enterprise | ${tagline}`,
-  CONTACT: `Contact | ${tagline}`,
-  PLAYGROUND: `Playground | ${tagline}`,
-}
-
-export const PageDescription = {
-  HOME: "API Platform engineered for scale.",
-  ABOUT: "API Platform engineered for scale.",
-  ENTERPRISE: "API Platform engineered for scale.",
-  CONTACT: "API Platform engineered for scale.",
-  PLAYGROUND: "Explore, test, and interact with your GraphQL APIs effortlessly",
 }
