+++ src/constants/titles.ts
@@ -1,0 +1,17 @@
+const tagline = "API Platform"
+
+export const PageTitle = {
+  HOME: tagline,
+  ABOUT: `About | ${tagline}`,
+  ENTERPRISE: `Enterprise | ${tagline}`,
+  CONTACT: `Contact | ${tagline}`,
+  PLAYGROUND: `Playground | ${tagline}`,
+}
+
+export const PageDescription = {
+  HOME: "API Platform engineered for scale.",
+  ABOUT: "API Platform engineered for scale.",
+  ENTERPRISE: "API Platform engineered for scale.",
+  CONTACT: "API Platform engineered for scale.",
+  PLAYGROUND: "Explore, test, and interact with your GraphQL APIs effortlessly",
+}
