+++ src/dep_pages/about.tsx
@@ -4,7 +4,7 @@ import Layout from "@theme/Layout"
 import {useLocation} from "@docusaurus/router"
 
 import AboutPage from "../components/about"
-import {PageDescription, PageTitle} from "../constants"
+import {PageDescription, PageTitle} from "../constants/titles"
 
 const About = (): JSX.Element => {
   const location = useLocation()
