+++ src/pages/playground.tsx
@@ -3,7 +3,7 @@ import ReactGA from "react-ga4"
 import Layout from "@theme/Layout"
 import PlaygroundPage from "../components/playground"
 import {useLocation} from "@docusaurus/router"
-import {PageDescription, PageTitle} from "../constants"
+import {PageDescription, PageTitle} from "../constants/titles"
 
 const Playground = () => {
   const location = useLocation()
